sem-chance code review.

+++ sem/biz1/sem-chance/src/main/java/com/bee32/sem/chance/dto/ChanceActionDto.java
@@ -103,7 +103,7 @@ public class ChanceActionDto
     protected void _parse(TextMap map)
             throws ParseException {
         plan = map.getString("plan") == "plan" ? true : false;
-        actor = new UserDto().ref(map.getLong("actorId"));
+        actor = new UserDto().ref(map.getInt("actorId"));
 
     }
